audiopolicy: do not set FM volume during device connection/disconnection
Volume inconsistencies are observed when a headset is
plugged in and out repeatedly while FM is active.

This is because the volume stored for stream music for the primary
output is first applied and then the FM volume for headset is applied.

Do not set volume as part of headset connection/disconnection to fix
the issue as it gets set as part of setOutputDevice anyway.

CRs-Fixed: 978396
Change-Id: Ibe0f3639ac4825330a710e86abc31d609136c924